remove CelTagService interface
